app/requirements.txt
paho-mqtt
requests
pyyaml
paho-mqtt
psycopg2-binary
dotenv
